Add Support for PDFs and Text Documents

+++ second-brain/brainly-backend/src/controllers/content.controller.ts
@@ -4,6 +4,14 @@ import { ApiResponse } from "../utils/ApiResponse";
 import { asyncHandler } from "../utils/asyncHandler";
 import { ApiError } from "../utils/ApiError";
 import { contentTypes } from "../utils/utilities";
+import { upload } from "../middlewares/multer.middleware";
+import {
+  extractPublicIdFromLink,
+  getThumbnailFromFile,
+  removeFromCloudinary,
+  uploadOnCloudinary,
+} from "../utils/cloudinary";
+import { UploadApiResponse } from "cloudinary";
 
 export const getAllContents = asyncHandler(async (req, res, next) => {
   const { q } = req.query;
@@ -30,20 +38,67 @@ export const getAllContents = asyncHandler(async (req, res, next) => {
   return res.status(200).json(new ApiResponse(200, contents));
 });
 
+export const handleOptionalFileUploadLocally = asyncHandler(
+  async (req, res, next) => {
+    // Middleware wrapper to handle file upload conditionally
+    upload.single("uploaded_file")(req, res, (err) => {
+      if (err) {
+        throw new ApiError(400, "File upload error");
+      }
+      console.log("No ERROR while uploading");
+      next();
+    });
+  }
+);
+
 export const addContent = asyncHandler(async (req, res, next) => {
   const contentSchema = z.object({
-    link: z.string().min(1, "link is required"),
-    type: z.enum(["image", "video", "article", "audio", "youtube", "tweet"]),
+    link: z.string().min(1, "link is required")?.optional(),
+    type: z.enum([
+      "image",
+      "video",
+      "article",
+      "audio",
+      "youtube",
+      "tweet",
+      "pdf",
+      "text",
+    ]),
     title: z.string().min(1, "Title is required"),
     tags: z.string().array().optional(), // or z.array(z.string());
   });
 
-  const { error, data } = contentSchema.safeParse(req.body);
+  const parsedBody = {
+    ...req.body,
+    tags: req.body?.tags ? JSON.parse(req.body.tags) : [], // Convert string to array
+  };
+
+  if (req.file) delete parsedBody.link;
+
+  const { error, data } = contentSchema.safeParse(parsedBody);
 
   if (error) throw new ApiError(400, "Please provide the required fields");
 
+  if (parsedBody.type === "pdf" || parsedBody.type === "text") {
+    const localFilePath = req.file?.path;
+    if (!localFilePath) throw new ApiError(400, "File is misssing");
+
+    const cloudinaryFile: UploadApiResponse | null = await uploadOnCloudinary(
+      localFilePath
+    );
+    // console.log(cloudinaryFile);
+
+    if (!cloudinaryFile)
+      throw new ApiError(400, "Error occured while uploading File to cloud");
+
+    // public_id is of form 'jd5hoprqerhzp1ocpyxa.txt'
+    // const thumbnail = await getThumbnailFromFile(cloudinaryFile.public_id);
+
+    parsedBody.link = cloudinaryFile.secure_url;
+  }
+
   await Content.create({
-    ...data,
+    ...parsedBody,
     user: req.user._id,
   });
   return res
@@ -53,6 +108,22 @@ export const addContent = asyncHandler(async (req, res, next) => {
 
 export const removeContent = asyncHandler(async (req, res, next) => {
   const { contentId } = req.params;
+
+  const content = await Content.findById(contentId);
+
+  if (!content) throw new ApiError(400, "Invalid Id");
+
+  if (content.type === "pdf" || content.type === "text") {
+    const public_id = extractPublicIdFromLink(content.link);
+
+    if (!public_id) throw new ApiError(400, "Invalid public Id");
+
+    // remove file from cloudinary
+    await removeFromCloudinary(
+      public_id,
+      content.type === "text" ? "raw" : "image"
+    );
+  }
 
   // Logged In user can delete his own content only
   await Content.deleteOne({
